(SERVER-2197) Use Server intermediate CA walk through
In https://github.com/puppetlabs/puppetserver/pull/1538 we added
dedicated documentation to Puppet Server outlining how to use it as an
intermediate CA. AFAICT, it is not linked anywhere from the site nav and
contains more, and more up to date information, than the information at
`config_ssl_external_ca`.

This deletes the section of `config_ssl_external_ca` and instead directs
readers to the Puppet Server documentation.