Get generated docs working with DITA
In the future we'll generate our reference docs in this repo, but
publish them along with the rest of our documentation so that the
reference docs work on the new docs site, and are included in our PDFs.

To this end, I've had to adjust the code to get the markdown output to
play nicely with the DITA-OT validation.

Also:
- Adds a workaround for the Facter 3 docs, so the build doesn't attempt
  to generate the CLI docs for Facter 3.
- Removes RONN and dumps the man pages into a codeblock instead. The man
  pages were causing issues, and I don't have time to mess around with
  this gem. Also, the docs should be just as (or more) readable if they
  keep their original man formatting.